factor-out embed ipc module
